requirements.txt
flask==3.0.0
pandas==2.1.4
plotly==5.17.0
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
flask-wtf==1.1.1
werkzeug==3.0.1
bcrypt==4.0.1
flask-limiter==3.5.0
cryptography==42.0.5
waitress==3.0.0
